rqt_reconfigure: Add ability to open clients from command line
This allows a user to list the names of dynamic reconfigure servers they wish
to connect to on startup. Unlike the original implementation in
reconfigure_gui, this makes rqt_reconfigure accept any number of clients, not
just one.

The servers can also be referenced using relative paths from the namespace
where rqt_reconfigure is started.