[rqt_image_view] Add a scroll area
When selecting a 1-to-1 pixel view, the previous implementation resized the
window to the full image size. This is inconvenient for images much larger than
the screen resolution. This patch places the RatioLayoutedFrame in a ScrollArea
so the window still fits on the screen and the user can pan on the full image.